+++ project/Component/nef/nef.swift
@@ -7,6 +7,7 @@ import NefJekyll
 import NefCarbon
 import NefModels
 
+
 // MARK: - Markdown <api>
 
 /// Renders content into Markdown files.
@@ -29,6 +30,7 @@ public func markdown(content: String, to outputPath: String,
                    success: success,
                    failure: failure)
 }
+
 
 // MARK: - Jekyll <api>
 
@@ -55,12 +57,12 @@ public func jekyll(content: String, to outputPath: String, permalink: String,
                  failure: failure)
 }
 
+
 // MARK: - Carbon <api>
 
 /// Renders a code selection into multiple Carbon images.
 ///
 /// - Precondition: this method must be invoked from main thread.
-/// - Postcondition: you should manage the output `NSWindow`.
 ///
 /// - Parameters:
 ///   - code: content to generate the snippet.
@@ -68,41 +70,7 @@ public func jekyll(content: String, to outputPath: String, permalink: String,
 ///   - outputPath: output where to render the snippets.
 ///   - success: callback to notify if everything goes well.
 ///   - failure: callback with information to notify if something goes wrong.
-///
-/// - Returns: `NSWindow` where Carbon snippet will be rendered.
 public func carbon(code: String,
-                   style: CarbonStyle,
-                   outputPath: String,
-                   success: @escaping () -> Void, failure: @escaping (String) -> Void) -> NSWindow {
-    guard Thread.isMainThread else {
-        fatalError("carbon(code:style:outputPath:success:failure:) should be invoked in main thread")
-    }
-    
-    let assembler = CarbonAssembler()
-    let window = assembler.resolveWindow()
-    
-    carbon(parentView: window.contentView!,
-           code: code,
-           style: style,
-           outputPath: outputPath,
-           success: success, failure: failure)
-    
-    return window
-}
-
-/// Renders a code selection into multiple Carbon images.
-///
-/// - Precondition: this method must be invoked from main thread.
-///
-/// - Parameters:
-///   - parentView: canvas view where to render Carbon image.
-///   - code: content to generate the snippet.
-///   - style: style to apply to exported code snippet.
-///   - outputPath: output where to render the snippets.
-///   - success: callback to notify if everything goes well.
-///   - failure: callback with information to notify if something goes wrong.
-public func carbon(parentView: NSView,
-                   code: String,
                    style: CarbonStyle,
                    outputPath: String,
                    success: @escaping () -> Void, failure: @escaping (String) -> Void) {
@@ -111,19 +79,16 @@ public func carbon(parentView: NSView,
     }
     
     let assembler = CarbonAssembler()
-    let carbonView = assembler.resolveCarbonView(frame: parentView.bounds)
-    let downloader = assembler.resolveCarbonDownloader(view: carbonView, multiFiles: false)
+    let window = assembler.resolveWindow()
+    let view = window.contentView!
+    let retainSuccess = { success(); _ = view }
+    let retainFailure = { (output: String) in failure(output); _ = view }
     
-    parentView.addSubview(carbonView)
-    
-    DispatchQueue(label: "nef-framework", qos: .userInitiated).async {
-        renderCarbon(downloader: downloader,
-                     code: "\(code)\n",
-                     style: style,
-                     outputPath: outputPath,
-                     success: success,
-                     failure: failure)
-    }
+    carbon(parentView: view,
+           code: code,
+           style: style,
+           outputPath: outputPath,
+           success: retainSuccess, failure: retainFailure)
 }
 
 /// Get an URL Request given a carbon configuration
